pyproject.toml
[project]
name = "fdm-halftone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fullcontrol",
    "pillow>=11.1.0",
]

[tool.uv.sources]
fullcontrol = { git = "https://github.com/FullControlXYZ/fullcontrol" }
